fix: router reworked, add cart price works
