--- src/components/RootLayout.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { Outlet } from 'react-router-dom';
-
-export default function RootLayout() {
-  return <Outlet />;
-}
