+++ src/pages/Basket.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export function Basket() {
+  return <p>Here should be a basket</p>;
+}
